small tweaks

+++ src/main/java/com/example/demo/repository/AccommodationRepo.java
@@ -3,6 +3,6 @@ package com.example.demo.repository;
 import com.example.demo.model.Accommodation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AccommodationRepo extends JpaRepository <Accommodation, Long> {
+public interface AccommodationRepo extends JpaRepository <Accommodation, Integer> {
 
 }
