+++ src/main/java/com/example/demo/repository/ActivityRepo.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.model.Activity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ActivityRepo extends JpaRepository<Activity, Long> {
+public interface ActivityRepo extends JpaRepository<Activity, Integer> {
 }
